Switch to hashbrown HashMap
The std-lib uses hashbrown internally but with a few different default
features. Noteworthy is that the hashbrown crate uses a faster hashing
algorithm: AHash instead of SipHash in the std-lib.
SipHash has the advantage of being better with collisions, which is not
a problem for us. Or at least AHash performs 25% faster taking us down
to 8.8s from 11.8s.
Also hashbrown has a "inline-more" feature that supposedly increases
speed in trade for compile time, but I can't measure any difference.
I'm also not sure if that feature is enabled in the std-lib.