Benchmark: Add warmup run
We are now fast enough that hyperfine is complaining about the first run
being a big outlier in terms of speed. I assume this is due to the os
caching loading the input file.
A single warmup round fixes this. Maybe I will look into ram-disks in
the future. This is what the original java version of BRC does.